requirements.txt
faiss-gpu
numpy
tqdm
opencv-contrib-python-headless
albumentations
torchsummary
timm
ptflops
